fix(config): WorkerSinkConfig.stats_interval_seconds is float
Match the rest of the codebase's interval-field type. Avoids a
ty-check widening warning when passed to WorkerSink(stats_interval_seconds: float)
during T4 scheduler wiring.

+++ src/pscanner/config.py
@@ -307,7 +307,7 @@ class WorkerSinkConfig(_Section):
     """
 
     velocity_maxsize: int = 4096
-    stats_interval_seconds: int = 60
+    stats_interval_seconds: float = 60.0
 
 
 class Config(BaseModel):
